refactor: extract useFADSideEffect hook
